backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
fastf1>=3.3.0
pandas>=2.1.0
numpy>=1.24.0
scipy>=1.11.0
httpx>=0.25.0
google-generativeai>=0.8.0

# Knowledge Base dependencies
chromadb>=0.4.22
pypdf>=4.0.0
langchain>=0.1.0
langchain-community>=0.0.20
sentence-transformers>=2.3.0
